persons-clubs completed

+++ app/services/skriesim/parsers/lookups/DisciplineTypesParser.scala
@@ -6,7 +6,7 @@ import services.skriesim.parsers.utils.ParserUtils
 object DisciplineTypesParser {
   def parse(html: String): Seq[IdName] = {
     ParserUtils.parseOptions(html, "Disciplīnu tips:").map {
-      case(id,name) => IdName(id.toInt,name)
+      case(id,name) => IdName(id.toLong,name)
     }
   }
 }
