+++ app/services/skriesim/parsers/lookups/RacesParser.scala
@@ -6,7 +6,7 @@ import services.skriesim.parsers.utils.ParserUtils
 object RacesParser {
   def parse(html: String): Seq[IdName] = {
     ParserUtils.parseOptions(html, "Sacensības nosaukums:").map {
-      case(id,name) => IdName(id.toInt,name)
+      case(id,name) => IdName(id.toLong,name)
     }
   }
 }
